+++ management/management.gen_test.go
@@ -2302,6 +2302,13 @@ func TestClient_GetRequirePushedAuthorizationRequests(tt *testing.T) {
 	c.GetRequirePushedAuthorizationRequests()
 }
 
+func TestClient_GetSessionTransfer(tt *testing.T) {
+	c := &Client{}
+	c.GetSessionTransfer()
+	c = nil
+	c.GetSessionTransfer()
+}
+
 func TestClient_GetSignedRequestObject(tt *testing.T) {
 	c := &Client{}
 	c.GetSignedRequestObject()
@@ -2946,6 +2953,16 @@ func TestClientRefreshToken_GetLeeway(tt *testing.T) {
 	c.GetLeeway()
 }
 
+func TestClientRefreshToken_GetPolicies(tt *testing.T) {
+	var zeroValue []ClientRefreshTokenPolicy
+	c := &ClientRefreshToken{Policies: &zeroValue}
+	c.GetPolicies()
+	c = &ClientRefreshToken{}
+	c.GetPolicies()
+	c = nil
+	c.GetPolicies()
+}
+
 func TestClientRefreshToken_GetRotationType(tt *testing.T) {
 	var zeroValue string
 	c := &ClientRefreshToken{RotationType: &zeroValue}
@@ -2969,6 +2986,34 @@ func TestClientRefreshToken_GetTokenLifetime(tt *testing.T) {
 func TestClientRefreshToken_String(t *testing.T) {
 	var rawJSON json.RawMessage
 	v := &ClientRefreshToken{}
+	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
+		t.Errorf("failed to produce a valid json")
+	}
+}
+
+func TestClientRefreshTokenPolicy_GetAudience(tt *testing.T) {
+	var zeroValue string
+	c := &ClientRefreshTokenPolicy{Audience: &zeroValue}
+	c.GetAudience()
+	c = &ClientRefreshTokenPolicy{}
+	c.GetAudience()
+	c = nil
+	c.GetAudience()
+}
+
+func TestClientRefreshTokenPolicy_GetScope(tt *testing.T) {
+	var zeroValue []string
+	c := &ClientRefreshTokenPolicy{Scope: &zeroValue}
+	c.GetScope()
+	c = &ClientRefreshTokenPolicy{}
+	c.GetScope()
+	c = nil
+	c.GetScope()
+}
+
+func TestClientRefreshTokenPolicy_String(t *testing.T) {
+	var rawJSON json.RawMessage
+	v := &ClientRefreshTokenPolicy{}
 	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
 		t.Errorf("failed to produce a valid json")
 	}
@@ -3389,6 +3434,16 @@ func TestConnectionOptions_GetPrecedence(tt *testing.T) {
 	c.GetPrecedence()
 	c = nil
 	c.GetPrecedence()
+}
+
+func TestConnectionOptions_GetRealmFallback(tt *testing.T) {
+	var zeroValue bool
+	c := &ConnectionOptions{RealmFallback: &zeroValue}
+	c.GetRealmFallback()
+	c = &ConnectionOptions{}
+	c.GetRealmFallback()
+	c = nil
+	c.GetRealmFallback()
 }
 
 func TestConnectionOptions_GetRequiresUsername(tt *testing.T) {
@@ -12888,6 +12943,14 @@ func TestResourceServerTokenEncryptionKey_String(t *testing.T) {
 	}
 }
 
+func TestRetryStrategy_String(t *testing.T) {
+	var rawJSON json.RawMessage
+	v := &RetryStrategy{}
+	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
+		t.Errorf("failed to produce a valid json")
+	}
+}
+
 func TestRMSClientAddon_GetURL(tt *testing.T) {
 	var zeroValue string
 	r := &RMSClientAddon{URL: &zeroValue}
@@ -14335,6 +14398,44 @@ func TestSessionDevice_GetLastUserAgent(tt *testing.T) {
 func TestSessionDevice_String(t *testing.T) {
 	var rawJSON json.RawMessage
 	v := &SessionDevice{}
+	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
+		t.Errorf("failed to produce a valid json")
+	}
+}
+
+func TestSessionTransfer_GetAllowedAuthenticationMethods(tt *testing.T) {
+	var zeroValue []string
+	s := &SessionTransfer{AllowedAuthenticationMethods: &zeroValue}
+	s.GetAllowedAuthenticationMethods()
+	s = &SessionTransfer{}
+	s.GetAllowedAuthenticationMethods()
+	s = nil
+	s.GetAllowedAuthenticationMethods()
+}
+
+func TestSessionTransfer_GetCanCreateSessionTransferToken(tt *testing.T) {
+	var zeroValue bool
+	s := &SessionTransfer{CanCreateSessionTransferToken: &zeroValue}
+	s.GetCanCreateSessionTransferToken()
+	s = &SessionTransfer{}
+	s.GetCanCreateSessionTransferToken()
+	s = nil
+	s.GetCanCreateSessionTransferToken()
+}
+
+func TestSessionTransfer_GetEnforceDeviceBinding(tt *testing.T) {
+	var zeroValue string
+	s := &SessionTransfer{EnforceDeviceBinding: &zeroValue}
+	s.GetEnforceDeviceBinding()
+	s = &SessionTransfer{}
+	s.GetEnforceDeviceBinding()
+	s = nil
+	s.GetEnforceDeviceBinding()
+}
+
+func TestSessionTransfer_String(t *testing.T) {
+	var rawJSON json.RawMessage
+	v := &SessionTransfer{}
 	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
 		t.Errorf("failed to produce a valid json")
 	}